Begun rewrite of kak-bundle with new design.
After contemplating a fair bit, I realized that cork.kak has a much better
design behind it, despite being quite a bit behind in features. By setting
up a file with the necessary files to load beforehand, and loading plugins
one at a time, it's possible to load all the .kak files in a plugin without
having to spawn a shell process, thus saving loading time dramatically. With
the old "load all plugins at once" approach, a shell process was necessary to
loop through all the plugins that needed to be loaded, which is a hit on loading
time.

With this commit, we are just as fast as cork.kak, while offering slightly
more features. Basically, while we aren't able to run from a shell, we do offer
the ability to grab plugin code from essentially anywhere, and run post-install
shell code for plugins that need a compilation step (kak-lsp and kak-rainbower
for example).

Next on the docket before merging into master is implementing the following:
- A means of registering a plugin without loading it (an equivalent to plug.kak's noload attribute)
- A means of selecting specific scripts from a plugin to load
- A means of updating plugins without deleting the directory in bundle_path
- A means of hard-resetting a plugin repo that won't otherwise update (basically bundle-force-update)

Of the above, the most difficult would probably be the picky load. While it would
be easy to do by spawning a shell process, that would increase load time and be
against the goals of this plugin manager, which is loading speed.