Redesign of bundle-install.
The benefits to this redesign of bundle-install are that instead
of needing to pass the installer of a plugin to the bundle-install
command to pick and choose which plugin to install, we can pass
the plugin names instead, which is much more user friendly. We also
now run the post-install hooks of each plugin inside of the plugin
directory, negating the need for an initial "cd $kak_opt_bundle_path/plugin"
step. Finally, only the post-install hooks of the plugins being installed
are run, meaning that kak-bundle won't waste time recompiling kak-lsp
when we only intended to install a colorscheme.