Improved handling of installers and post-install hooks.
The primary issue with the previous approach of storing the plugin
installers and post-install hooks in a file is that bundle-clean
couldn't remove these files - otherwise, bundle-install and bundle-update
couldn't retrieve the installers and post-install hooks before the next
Kakoune restart. This is obviously an issue, because bundle-clean should
be expected to remove all traces of the cleaned plugin from the plugin
install directory.

For this, we needed a way to store dictionaries of values with spaces in
a Kakoune option. Unfortunately, there's no easy way to do this that
I know of, because str-list and str-to-str-map use spaces as separators,
which is an issue because installers and post-install hooks often
contain spaces. To resolve this, we use a hack where the "dictionary" is
essentially a str option with the keys and values separated by the 🦀
character. Obviously, this is not a perfect solution, because it would
break if an installer or post-install hook contained the 🦀 character,
but it should be good enough for 99.99% of cases.