Added bundle_parallel option to support parallel install/updates.
The main challenge with implementing parallelism with commands
that generate output is to avoid mangling the output of multiple
commands and creating a confusing mess. How the previous version
of kak-bundle resolved this was to avoid using a FIFO, and instead
to periodically read "output files" and write their contents to a
buffer. However, this had the side effect of not easily being able
to render the output of the post-install hooks in real time, requiring
the awkward step of opening up a FIFO buffer.

How we're doing it this time is to employ a trick where we write the
output of a command to a file, but don't show the output until the
command finishes, at which point all the command's output is shown
at once. While not a particularly elegant solution, it works for the
simpler FIFO-based approach we are using, and keeps the kak-bundle
script simple.