Added bundle-after-install hook; updated README.
The main motivation behind this change was that a valid use case for
a plugin manager is to update its plugins from the command line - say,
during a systemwide upgrade process. We can accomplish this by triggering
a hook after install/update to quit Kakoune and take the user back to the
command line.