Removed kak-bundle.sh.
This worked by accident with the simplified UI, since the shell
script was only called once during install, but the old UI that
I reverted back to needed to call it multiple times, which was a
problem if kak-bundle was one of the plugins being reinstalled,
since kak-bundle.sh would no longer exist in that case. Having
all the shell code in a hidden Kakoune option bypasses this
issue, and I'm honestly not sure why I changed it in the first
place.